Fix type

+++ crates/cargo-contract/src/cmd/config.rs
@@ -61,9 +61,9 @@ impl Config for Ecdsachain {
 impl Environment for Ecdsachain {
     const NATIVE_TO_ETH_RATIO: u32 =
         <DefaultEnvironment as Environment>::NATIVE_TO_ETH_RATIO;
-    const TRUST_BACKED_ASSETS_PRECOMPILE_INDEX: u32 =
+    const TRUST_BACKED_ASSETS_PRECOMPILE_INDEX: u16 =
         <DefaultEnvironment as Environment>::TRUST_BACKED_ASSETS_PRECOMPILE_INDEX;
-    const POOL_ASSETS_PRECOMPILE_INDEX: u32 =
+    const POOL_ASSETS_PRECOMPILE_INDEX: u16 =
         <DefaultEnvironment as Environment>::POOL_ASSETS_PRECOMPILE_INDEX;
 
     type AccountId = <DefaultEnvironment as Environment>::AccountId;
@@ -99,9 +99,9 @@ impl Config for Substrate {
 impl Environment for Substrate {
     const NATIVE_TO_ETH_RATIO: u32 =
         <DefaultEnvironment as Environment>::NATIVE_TO_ETH_RATIO;
-    const TRUST_BACKED_ASSETS_PRECOMPILE_INDEX: u32 =
+    const TRUST_BACKED_ASSETS_PRECOMPILE_INDEX: u16 =
         <DefaultEnvironment as Environment>::TRUST_BACKED_ASSETS_PRECOMPILE_INDEX;
-    const POOL_ASSETS_PRECOMPILE_INDEX: u32 =
+    const POOL_ASSETS_PRECOMPILE_INDEX: u16 =
         <DefaultEnvironment as Environment>::POOL_ASSETS_PRECOMPILE_INDEX;
 
     type AccountId = <DefaultEnvironment as Environment>::AccountId;
@@ -134,9 +134,9 @@ impl Config for Polkadot {
 impl Environment for Polkadot {
     const NATIVE_TO_ETH_RATIO: u32 =
         <DefaultEnvironment as Environment>::NATIVE_TO_ETH_RATIO;
-    const TRUST_BACKED_ASSETS_PRECOMPILE_INDEX: u32 =
+    const TRUST_BACKED_ASSETS_PRECOMPILE_INDEX: u16 =
         <DefaultEnvironment as Environment>::TRUST_BACKED_ASSETS_PRECOMPILE_INDEX;
-    const POOL_ASSETS_PRECOMPILE_INDEX: u32 =
+    const POOL_ASSETS_PRECOMPILE_INDEX: u16 =
         <DefaultEnvironment as Environment>::POOL_ASSETS_PRECOMPILE_INDEX;
 
     type AccountId = <DefaultEnvironment as Environment>::AccountId;
